Moving OHLCV data wrangling to its own module
