+++ tradingstrategy/utils/groupeduniverse.py
@@ -31,6 +31,8 @@ from tradingstrategy.utils.forward_fill import forward_fill
 from tradingstrategy.utils.forward_fill import forward_fill as _forward_fill
 from tradingstrategy.utils.time import assert_compatible_timestamp, ZERO_TIMEDELTA, naive_utcnow
 
+from .wrangle import fix_bad_wicks, filter_bad_wicks, remove_zero_candles
+
 logger = logging.getLogger(__name__)
 
 
@@ -1079,127 +1081,3 @@ def resample_price_series(
     series = series.resample(resample_freq).agg(func)
     return series
 
-
-def fix_bad_wicks(
-    df: pd.DataFrame,
-    threshold=(0.1, 1.9),
-    too_slow_threshold=15,
-    verbose=False,
-    bad_open_close_threshold: float | None=3.0,
-) -> pd.DataFrame:
-    """Correct out bad high/low values in OHLC data.
-
-    Applicable for both :term:`OHLCV` price feeds and liquidity feeds.
-
-    On :term:`Uniswap` v2 and compatibles, Bad wicks are caused by e.g. very large flash loan, oracle price manipulation attacks,
-    and misbheaving bots.
-
-    This function removes bad high/low values and sets them to open/close if they seem to be wildly out of sample.
-
-    :param threshold:
-        How many pct % wicks are allowed through.
-
-        Tuple (low threshold, high threshold) relative to close.
-
-        Default to 50%. A high wick cannot be more than 50% of close.
-
-    :param too_slow_threshold:
-        Complain if this takes too long
-
-    :param bad_open_close_threshold:
-        How many X open must be above the high to be considered a broken data point.
-
-        The open price will be replaced with high price.
-
-        Do not set for liquidity processing.
-
-
-    :param verbose:
-        Make some debug logging when using the function for manual data diagnostics.
-
-    """
-
-    start = naive_utcnow()
-
-    if len(df) == 0:
-        return df
-
-
-    # Optimised with np.where()
-    # https://stackoverflow.com/a/65729035/315168
-    if threshold is not None:
-        df["high"] = np.where(df["high"] > df["close"] * threshold[1], df["close"], df["high"])
-        df["low"] = np.where(df["low"] < df["close"] * threshold[0], df["close"], df["low"])
-
-    # For manual diagnostics tracking down bad trading pair data
-    if verbose and bad_open_close_threshold:
-        bad_opens = df[df["open"] > df["high"] * bad_open_close_threshold]
-        for idx, row in islice(bad_opens.iterrows(), 10):
-            logger.warning(
-                "Pair id %d, timestamp: %s, open: %s, high: %s, buy volume: %s sell volume: %s, volume: %s",
-                row.pair_id,
-                row.timestamp,
-                row.open,
-                row.high,
-                row.get("buy_volume"),
-                row.get("sell_volume"),
-                row.get("volume"),
-            )
-        logger.warn("Total %d bad open price entries detected", len(bad_opens))
-
-    # Issues in open price values with data point - open cannot be higher than high.
-    # Not strickly "wicks" but we fix all data while we are at it.
-    if bad_open_close_threshold:
-        df["open"] = np.where(df["open"] > df["high"] * bad_open_close_threshold, df["high"], df["open"])
-        df["close"] = np.where(df["close"] > df["high"] * bad_open_close_threshold, df["high"], df["close"])
-
-    duration = naive_utcnow() - start
-
-    if duration > datetime.timedelta(seconds=too_slow_threshold):
-        logger.warning("Very slow fix_bad_wicks(): %s", duration)
-
-    # The following code chokes
-    # mask = (df["high"] > df["close"] * (1+threshold)) | (df["low"] < df["close"] * threshold)
-    #df.loc[mask, "high"] = df["close"]
-    #df.loc[mask, "low"] = df["close"]
-    #df.loc[mask, "wick_filtered"] = True
-    return df
-
-
-def filter_bad_wicks(df: pd.DataFrame, threshold=(0.1, 1.9)) -> pd.DataFrame:
-    """Mark the bad wicks.
-
-    On :term:`Uniswap` v2 and compatibles, Bad wicks are caused by e.g. very large flash loan, oracle price manipulation attacks,
-    and misbheaving bots.
-
-    This function removes bad high/low values and sets them to open/close if they seem to be wildly out of sample.
-
-    :param threshold:
-        How many pct % wicks are allowed through as (low, high) tuple.
-
-        This is a tuple (low threshold, high threshold).
-
-        If low < close * threshold[0] ignore the value.
-
-        If high > close * threshold[0] ignore the value.
-
-    """
-
-    df_matches = df.loc[
-        (df["high"] > df["close"] * threshold[1]) | (df["low"] < df["close"] * threshold[0])
-    ]
-
-    return df_matches
-
-def remove_zero_candles(
-    df: pd.DataFrame,
-) -> pd.DataFrame:
-    """Remove any candle that has a zero value for OHLC
-    
-    :param df: Dataframe that may contain zero candles
-    :return: pd.Dataframe
-    """
-    if len(df) > 0:
-        filtered_df = df[(df['open'] != 0) & (df['high'] != 0) & (df['low'] != 0) & (df['close'] != 0)]
-        return filtered_df
-    return df
